Explicitly cast return value of read(2) to "int"
This fixes the following gcc warning, emitted by Xcode:

src/ngircd/sighandlers.c: In function 'Signal_Callback':
src/ngircd/sighandlers.c:239: warning: implicit conversion shortens 64-bit value into a 32-bit value

+++ src/ngircd/sighandlers.c
@@ -236,7 +236,7 @@ Signal_Callback(int fd, short UNUSED what)
 	(void) what;
 
 	do {
-		ret = read(fd, &sig, sizeof(sig));
+		ret = (int)read(fd, &sig, sizeof(sig));
 		if (ret == sizeof(int))
 			Signal_Handler_BH(sig);
 	} while (ret == sizeof(int));
